WIP all non-test files building
