+++ pkg/chain/tbtc.go
@@ -17,8 +17,7 @@ type TBTCHandle interface {
 // Deposit is an interface that provides ability to interact
 // with Deposit contracts.
 type Deposit interface {
-	// KeepAddress returns the underlying keep address for the
-	// provided deposit.
+	// Keep returns the underlying keep for the provided deposit.
 	Keep(depositAddress string) (BondedECDSAKeepHandle, error)
 
 	// RetrieveSignerPubkey retrieves the signer public key for the
